Use the fixture_auto_set_env_vars fixture for the entire fastapi test suite
