+++ tests/nat/front_ends/fastapi/conftest.py
@@ -1,0 +1,24 @@
+# SPDX-FileCopyrightText: Copyright (c) 2025, NVIDIA CORPORATION & AFFILIATES. All rights reserved.
+# SPDX-License-Identifier: Apache-2.0
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+# http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+
+import pytest_asyncio
+
+
+@pytest_asyncio.fixture(name="auto_set_env_vars", autouse=True)
+async def fixture_auto_set_env_vars(setup_db,
+                                    set_nat_config_file_env_var,
+                                    set_nat_dask_scheduler_env_var,
+                                    set_nat_job_store_db_url_env_var):
+    return
